chore(ui): client-side renders table column headings
